requirements.txt
streamlit>=1.22.0
google-generative-ai>=0.3.0
easyocr>=1.6.2
opencv-python-headless>=4.7.0.72
pdf2image>=1.16.0
pillow>=9.5.0
pdfplumber>=0.9.0
numpy>=1.23.5